Refactor pay logic, remove ServletUtils, update IP utils

+++ ruoyi-pay/src/main/java/com/ruoyi/pay/service/app/PayAppServiceImpl.java
@@ -48,7 +48,7 @@ public class PayAppServiceImpl implements PayAppService {
     PayAppDO app = PayAppConvert.INSTANCE.convert(createReqVO);
     appMapper.insert(app);
     // 返回
-    return app.getId();
+    return app.getId().intValue();
   }
 
   @Override
